Add link to docs to README
While the GH repo's site is the docs page, that is off-center (and not rendered when viewing the repository via a cloned local view). I had to be explicitly told the docs site existed due to not noticing that location (nor any other which may or may not exist). Ideally, the README includes the link *somewhere*. This is a location with no changes to layout.