fix: Fix fake_exponential for use with large inputs
The previous implementation would return incorrect values for large
numerators and in those cases differed from the implementation in geth.
This was caused internal calculations overflowing.

Now arbitrarily sized values are used for the calculation, however if
the result does not fit in a u128 then the function will panic instead
of returning an invalid value.